requirements.txt
httpx
tldextract
rapidfuzz
beautifulsoup4
kivy
pyjnius
certifi
